move ministry contacts into own constants file
