+++ services/minespace-web/src/constants/ministryContacts.js
@@ -1,0 +1,16 @@
+// Ministry Contacts
+export const MINISTRY_CONTACTS = {
+  REGIONAL_HEALTH_AND_SAFETY: {
+    NAME: "John Doe",
+    PHONE: "(250) 479-2546",
+    EMAIL: "john.doe@gov.bc.ca",
+  },
+  REGIONAL_MINE_GENERAL: {
+    EMAIL: "inforegionalmines@gov.bc.ca",
+  },
+  MAJOR_MINE_GENERAL: {
+    EMAIL: "infomajormines@gov.bc.ca",
+  },
+};
+
+export default MINISTRY_CONTACTS;
